requirements.txt
# ============================================
# CORE DJANGO & API
# ============================================
Django==6.0.1
djangorestframework==3.15.2
django-filter==24.3
django-cors-headers==4.4.0

# ============================================
# DATABASE & PRODUCTION SERVER
# ============================================
psycopg2-binary==2.9.10
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0

# ============================================
# SECURITY & AUTHENTICATION
# ============================================
PyJWT==2.8.0
django-ratelimit==4.1.0

# ============================================
# CACHING (REDIS)
# ============================================
django-redis==5.4.0
redis==5.0.1

# ============================================
# UTILITIES
# ============================================
requests==2.31.0  # Often needed for health checks or external APIs